backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
alembic==1.13.0
redis==5.0.1
langchain==0.1.5
openai==1.6.1
websockets==12.0
python-multipart==0.0.6
alembic==1.13.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
requests==2.31.0
tiktoken==0.5.1
email-validator==2.1.0.post1